Add files via upload
This is the first version of the neural network. Sample data is loaded into an RDD with 2304 input variables (1024 from the bit vector and 1280 from the esm embedding f the protein). The neural network just applies a bunch of linear transformations with a leakyRelu between each layer. It technically works but it functions super slowly and with a batch size of 32 there are still 2965 batches per epoch, making it run very slow.